Fix cyclic dependency by moving everything to the same package
The rationale is to make it work, then fix it if the approach proves
correct.
